requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pydantic==2.9.2
aiofiles==24.1.0
bcrypt==4.0.1
psycopg2-binary==2.9.10
neo4j==5.26.0
sentence-transformers>=3.0.0
transformers==4.57.6
torch>=2.0.0
Pillow>=10.0.0
pdf2image>=1.17.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.28
langchain-openai>=0.1.25
langchain-core>=0.2.43
einops
timm